Use standardized spectral residuals
